refactor: User -> user 디렉토리 이름 변경
패키지명은 소문자를 사용하는 관습에 따라 기존의 디렉토리명을 User -> user 변경했습니다.

하위 디렉토리인 Domain, DTO, Service, Repository도 모두 소문자로 디렉토리명을 변경했습니다.
