+++ movie-book/src/main/java/hello/moviebook/User/domain/Auth.java
@@ -1,4 +1,4 @@
-package hello.moviebook.User;
+package hello.moviebook.user.domain;
 
 public enum Auth {
     USER, ADMIN
